fix(#517): decompose 711-line config.go into three focused files
Split config.go by responsibility:
- config.go: BuildCaddyConfig, TLS validation and app builders, HTTP redirect server
- config_routes.go: route builders (health, ready, metrics, admin, kratos, catch-all)
- config_handlers.go: handler builders (user header strip, security headers)

Pure refactor — no behaviour change. All tests pass.
